change how location of EdgeMarkers are calculated
the original just used the x or y coordinates of target.
this is close to correct. For examle:
if you have EdgeMarker for a point very far away to left and bit more north.
edge maker will be placed in the corner. because bothe x and y are from outside screen reach.
this gives the impression that target is in direction of that corner.
but in reality it's to your left(and a small bit up)

the new implementation place marker on the intersection of line from
center of creen to target and the edge

its has small fault:
is place the marker with small offset from edge of  the screen
this offset is the height of icon for top/bottom and with for
left/right edge
this gives incorrect result if icon is not square and rotated